+++ Exercism/c/sum-of-multiples/src/sum_of_multiples.h
@@ -1,0 +1,6 @@
+#ifndef SUM_OF_MULTIPLES_H
+#define SUM_OF_MULTIPLES_H
+
+long sum_of_multiples(const unsigned int*, int, int);
+
+#endif